+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2023-10-22
+
+### Fixed
+
+- Open external docs: Use `xdg-open` or `open` (MacOS) by default
+  and fall back to `netrw`.
+  Remove redundant URL encoding.
+
 ## [2.1.0] - 2023-10-22
 
 ### Added